implement emitter
