+++ xray/segment_processor.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/xray"
+	"google.golang.org/appengine/log"
 	"sync"
 )
 
@@ -21,21 +22,21 @@ func NewSegmentProcessor(awsConfig aws.Config, batchSize int) *SegmentProcessor 
 	return &processor
 }
 
-func (p *SegmentProcessor) transform(batch *[]string) *[][]string {
+func (p *SegmentProcessor) transform(batch []string) [][]string {
 	batches := make([][]string, 0)
 	curBatch := make([]string, 0)
-	for idx, s := range *batch {
+	for idx, s := range batch {
 		if idx > 0 && idx % p.batchSize == 0 {
 			batches = append(batches, curBatch)
 			curBatch = make([]string, 0)
 		}
 		curBatch = append(curBatch, s)
 	}
-	return &batches
+	return batches
 }
 
-func (p *SegmentProcessor) Transmit(ctx *context.Context, batch *[]string) {
-	batches := *(p.transform(batch))
+func (p *SegmentProcessor) Transmit(ctx *context.Context, batch []string) {
+	batches := p.transform(batch)
 	numProcessors := len(batches)
 	var group sync.WaitGroup
 	group.Add(numProcessors)
@@ -43,4 +44,16 @@ func (p *SegmentProcessor) Transmit(ctx *context.Context, batch *[]string) {
 		go EmitSegments(p.XRayClient, ctx, &(batches[i]), group)
 	}
 	group.Wait()
+}
+
+// transmit a single batch to XRay
+func EmitSegments(client *xray.Client, ctx *context.Context, batch *[]string, group sync.WaitGroup) {
+	req := &xray.PutTraceSegmentsInput{TraceSegmentDocuments: *batch}
+	putSegReq := client.PutTraceSegmentsRequest(req)
+	_, err := putSegReq.Send(*ctx)
+	if err != nil {
+		log.Errorf(*ctx, err.Error())
+	}
+
+	group.Done()
 }
